refactore a bit
Implement MSV3 purchase order #3411
